always alloc to size
Currently, calling GetGrown(X) will return a slice of cap=X if it
allocates anew. Later, if that same slice is used by a GetGrown(Y)
call, it may need to be grown to fit Y.

That can cause a lot of extra copying and allocation.

To help that, have GetGrown/GetFilled always alloc to the relevant
bucket size. If GetGrown(X) and GetGrown(Y) are served by the same
bucket the slices they return will have cap equal to that bucket's
size.